CT-2257 Refactor notes and messages controllers to use notable concern
